Make limit and offset queries work with relevance sorting
Previously calling `.records.limit(n)` would result in not getting the
most relevant records because MySQL would do the limit before the
in-memory sorting that we do here. Now we do the limit and offset on the
ids which are already in order of relevance and then we nullify them
through the relation to not do them in SQL at all.