We need to use 0 instead of nil.
The will_paginate gem returns the offset value if you pass in nil which
is overriding the default behaviour of ActiveRecord. This is not what we
want because we want to return the relation (or in this case the
relation like object). By using 0 we fix this problem.

Barbara Liskov would be upset.